requirements.txt
selenium==4.27.1
webdriver_manager==4.0.2
retry==0.9.2
python-dotenv==1.0.1
ipykernel== 6.29.5
pandas==2.2.3
ollama==0.5.1